wip(tama): checkpoint job_0mtmii5gt00bkgkb5
